ciifen_digitalizador_backend-main/ciifen_digitalizador_backend-main/requirements/base.txt
asgiref==3.7.2
Django==3.2.19
django-cors-headers==3.7.0
django-environ==0.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.3
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
Pillow==10.0.0
opencv-python~=4.8.0.74
numpy~=1.25.1
matplotlib~=3.7.2
scipy~=1.11.1